Reorganised so tests for each topology can be run separately.  Fixed timeout so UDP tests that fail to transmit any iperf data do not kill the test.  Added zero data for when no ipref data is generated.

+++ tve_ipv6_and_multicast/runTestsRepeated.bash
@@ -1,11 +1,13 @@
 #!/bin/bash
 d=`dirname $0`
 basedir=`cd ${d}; pwd`
-for ((i=1; i<=$2; i++)); do
-	sudo killall -9 iperf3
+for ((i=1; i<=$3; i++)); do
+	if [ `ps aux | grep iperf3 | grep -v grep | wc -l` -gt 0 ]; then
+		sudo killall -9 iperf3
+	fi
 	echo "===== Iteration: $i ====="
-	sudo `which python` ${basedir}/testsets/$1.py
-	if [ $i -ne $2 ]; then
+	sudo `which python` ${basedir}/testsets/${1}/test${2}.py
+	if [ $i -ne $3 ]; then
 		sleep 30
 	fi
 done
